snake game app.
In this  game player has to monitor the direction for snake to get its
egg, as it consumes food its length will be increased. The player loses
when the snake runs into the screen border or if it runs over itself

If the player has to quit this game then he can use "q" to quit the
game. after quiting the game, total score will be on console. the score
is incremented, when the snake eats its food.

What is the impact of the change?
.....